create working paginations
1/ create all buttons paginations
2/ create basic CSS
