+++ src/partials/paginations.html
@@ -1,4 +1,50 @@
+<!-- Paginations -->
 <div>
-    <ul class="paginations-slider"></ul>
+  <ul class="paginations-slider"></ul>
 </div>
-<div class="paginations-container"></div>
+<!-- buttons paginations -->
+<div class="paginations">
+  <div class="paginations__container paginations__container-back">
+    <button
+      type="button"
+      name="startButton"
+      class="paginations__btn paginations__btn--back"
+    >
+      <svg class="paginations__icon" width="24" height="24">
+        <use href="./images/symbol-defs.svg#icon-double-small-left"></use>
+      </svg>
+    </button>
+    <button
+      type="button"
+      name="previousButton"
+      class="paginations__btn paginations__btn--back"
+    >
+      <svg class="paginations__icon" width="24" height="24">
+        <use href="./images/symbol-defs.svg#icon-small-left"></use>
+      </svg>
+    </button>
+    <!-- container for dynamic pagination page markup -->
+  </div>
+  <div class="paginations__container paginations__container-pages"></div>
+  <!-- buttons paginations -->
+  <div class="paginations__container paginations__container-end">
+    <button
+      type="button"
+      name="nextButton"
+      class="paginations__btn paginations__btn--end"
+    >
+      <svg class="paginations__icon" width="24" height="24">
+        <use href="./images/symbol-defs.svg#icon-small-right"></use>
+      </svg>
+    </button>
+    <button
+      type="button"
+      name="endButton"
+      class="paginations__btn paginations__btn--end"
+    >
+      <svg class="paginations__icon" width="24" height="24">
+        <use href="./images/symbol-defs.svg#icon-double-small-right"></use>
+      </svg>
+    </button>
+  </div>
+</div>
