--- src/templates/btnPaginations.hbs
@@ -1,4 +1,0 @@
-<button></button>
-<button></button>
-<button></button>
-<button></button>
