--- src/templates/paginations.hbs
@@ -1,10 +1,0 @@
-   
-
-    <li class="paginations-slider__item">
-        <p>name: <b>{{name}}</b></p>
-        <p>age: {{age}}</p>
-        <p>city: {{address.city}}</p>
-        <p>state: {{address.state}}</p>
-    </li>
-  
-
